+++ addon/components/flexberry-menu.js
@@ -32,6 +32,17 @@ import FlexberryBaseComponent from './flexberry-base-component';
         icon: 'trash icon',
         title: 'Delete',
       }],
+    }, {
+      title: 'Item',
+      buttons: [{
+        title: 'edit',
+        buttonClasses: 'icon',
+        iconClass: 'edit icon',
+        disabled: false,
+        buttonAction: () => {
+          window.alert('clicked edit button');
+        }
+      }]
     }],
     ...
     actions: {
@@ -87,12 +98,12 @@ export default FlexberryBaseComponent.extend({
   items: null,
 
   /**
-    See [Semantic UI API](https://semantic-ui.com/modules/dropdown.html#/settings).
+    Settings for the dropdown, see [Semantic UI API](https://semantic-ui.com/modules/dropdown.html#/settings) for more info.
 
     @property settings
     @type Object
   */
- settings: undefined,
+  settings: undefined,
 
   /**
     Array CSS class names.
@@ -227,6 +238,11 @@ export default FlexberryBaseComponent.extend({
   _onClickHandler(e) {
     // Find clicked menu item element.
     let itemElement = $(e.target);
+
+    if (itemElement.parent().is('button') || itemElement.is('button')) {
+      return;
+    }
+
     if (!itemElement.hasClass('flexberry-menuitem')) {
       itemElement = itemElement.parents('.flexberry-menuitem');
     }