+++ addon/components/mobile/object-list-view.js
@@ -16,6 +16,7 @@ export default ObjectListViewComponent.extend({
   /**
     Flag indicates whether visible selected menu for mobile.
 
+    @private
     @property _selectedMobileMenu
     @type Boolean
     @readOnly
@@ -27,6 +28,7 @@ export default ObjectListViewComponent.extend({
   /**
     Count selected row for mobile menu.
 
+    @private
     @property _selectedCountMobileMenu
     @type Number
     @readOnly
@@ -93,12 +95,67 @@ export default ObjectListViewComponent.extend({
     });
   },
 
+  /**
+    @private
+    @property _checkRowsSettingsItems
+    @readOnly
+  */
+  _checkRowsSettingsItems: computed(
+    'i18n.locale',
+    'userSettingsService.isUserSettingsServiceEnabled',
+    'readonly',
+    'allSelect',
+    'allSelectAtPage',
+    function() {
+      let i18n = this.get('i18n');
+      let readonly = this.get('readonly');
+      let allSelect = this.get('allSelect');
+
+      let rootItem = {
+        icon: 'dropdown icon',
+        iconAlignment: 'right',
+        title: '',
+        items: [],
+        localeKey: ''
+      };
+
+      let isUncheckAllAtPage = this.get('allSelectAtPage');
+      let checkAllAtPageTitle = isUncheckAllAtPage ? i18n.t('components.olv-toolbar.uncheck-all-at-page-button-text') :
+      i18n.t('components.olv-toolbar.check-all-at-page-button-text');
+      let checkAllAtPageTitleKey = isUncheckAllAtPage ? 'components.olv-toolbar.uncheck-all-at-page-button-text' :
+      'components.olv-toolbar.check-all-at-page-button-text';
+
+      let checkAllTitle = allSelect ? i18n.t('components.olv-toolbar.uncheck-all-button-text') :
+      i18n.t('components.olv-toolbar.check-all-button-text');
+      let checkAllTitleKey = allSelect ? 'components.olv-toolbar.uncheck-all-button-text' :
+      'components.olv-toolbar.check-all-button-text';
+
+      if (!readonly) {
+        if (!allSelect) {
+          rootItem.items.push({
+            title: checkAllAtPageTitle,
+            localeKey: checkAllAtPageTitleKey
+          });
+        }
+
+        let classNames = this.get('classNames');
+        if (classNames && classNames.indexOf('groupedit-container') === -1) {
+          rootItem.items.push({
+            title: checkAllTitle,
+            localeKey: checkAllTitleKey
+          });
+        }
+      }
+
+      return this.get('userSettingsService').isUserSettingsServiceEnabled ? [rootItem] : [];
+    }
+  ),
+
   actions: {
     /**
       Delete selected rows.
 
       @method actions.deleteSelectedRow
-      @public
     */
     deleteSelectedRow() {
       let confirmDeleteRows = this.get('confirmDeleteRows');
@@ -127,29 +184,16 @@ export default ObjectListViewComponent.extend({
       Clear selected rows.
 
       @method actions.clearSelectedRecords
-      @public
     */
     clearSelectedRecords() {
       if (this.get('allSelect')) {
         this.send('checkAll');
-      }
-
-      let componentName = this.get('componentName');
-      let contentWithKeys = this.get('contentWithKeys');
-      let selectedRecords = this.get('selectedRecords');
-      let selectedRows = contentWithKeys.filterBy('selected', true);
-      for (let i = 0; i < selectedRows.length; i++) {
-        let recordWithKey = selectedRows[i];
-        let selectedRow = this._getRowByKey(recordWithKey.key);
-
-        if (selectedRow.hasClass('active')) {
-          selectedRow.removeClass('active');
-        }
-
-        selectedRecords.removeObject(recordWithKey.data);
-        recordWithKey.set('selected', false);
-
-        this.get('objectlistviewEventsService').rowSelectedTrigger(componentName, recordWithKey.data, selectedRecords.length, false, recordWithKey);
+      } else {
+        const contentWithKeys = this.get('contentWithKeys');
+        this.get('selectedRecords').map((record) => contentWithKeys.findBy('data', record)).forEach((modelWithKey) => {
+          modelWithKey.set('selected', false);
+          this.send('selectRow', modelWithKey, { checked: false });
+        });
       }
     },
   }