+++ addon/locales/en/translations.js
@@ -114,7 +114,6 @@ export default {
         }
       }
     },
-
     'flexberry-objectlistview': {
       'placeholder': 'There is no data',
       'showing-entries': {
@@ -122,6 +121,7 @@ export default {
         'of': ' of ',
         'entries': ' entries'
       },
+      'without-sorting': 'Without sorting',
       'search-page-placeholder': '№ page',
       'search-button-text': 'Go to page'
     },
@@ -267,12 +267,10 @@ export default {
     },
 
     'filters-dialog-content': {
-      'title': 'Сolumns filtering',
-      'column-name': 'Filter column',
-      'column-condition': 'Condition',
-      'column-value': 'Value',
-      'clear': 'Clear filters',
-      'apply': 'Apply filters'
+      'clear-this-filter': 'Reset this filter',
+      'title': 'Configuring record filtering',
+      'clear': 'Reset filter',
+      'apply': 'Apply'
     },
 
     'form-load-time-tracker': {